Cleanup web3 mod move to web3-react dir
